Alternative config bool_to_env.
Make this a little more "pythony" style.
Just attempt to convert var to bool, and reraise exception if fails.
Speciying the defaults as 0 or 1 might make it clear what the user is supposed to provide.

Only have single underscore for "private" methods. Double leading underscore is a convention reserved for:
https://www.python.org/dev/peps/pep-0008/
```
_double_leading_underscore : when naming a class attribute, invokes name mangling (inside class FooBar, __boo becomes _FooBar__boo ; see below).

__double_leading_and_trailing_underscore__ : "magic" objects or attributes that live in user-controlled namespaces. E.g. __init__ , __import__ or __file__ . Never invent such names; only use them as documented.
```